Remove more `BuiltinLintDiag` in `rustc_resolve`
